Add parser utility and tests
